refactor: `calcYearIntervals` to utils
